feat: improve quick install and allow it to be run in current dir

+++ packages/create-kit-docs/bin/create-kit-docs.js
@@ -20,53 +20,72 @@ async function main() {
   const rootDirName = argv._[0];
 
   if (!rootDirName) {
-    console.log(
-      kleur.red(
-        `\n${kleur.bold('[kit-docs]: missing directory name!')}\n\n${kleur.white(
-          `Example: ${kleur.cyan('`npm init @svelteness/kit-docs mydocs`')}`,
-        )}\n`,
-      ),
-    );
-    return;
+    const { ok } = await useCurrentDirPrompt();
+    if (!ok) return;
   }
 
-  let targetDir = path.resolve(__cwd, rootDirName);
-
-  if (fs.existsSync(targetDir)) {
-    console.log();
-    const { overwrite } = await overwritePrompt();
-    if (!overwrite) return;
-    await emptyDir(targetDir);
-  }
+  let targetDir = path.resolve(__cwd, rootDirName ?? '');
 
   const version = JSON.parse(fs.readFileSync(__pkg).toString()).version;
 
-  console.log();
-  const { defaultTheme } = await defaultThemePrompt();
+  if (rootDirName) console.log();
+  const { ok: defaultTheme } = await defaultThemePrompt();
 
-  try {
-    const child = spawn('npm', ['init', 'svelte@next', rootDirName], {
-      stdio: 'inherit',
-    });
+  const isTargetDirEmpty = isDirEmpty(targetDir);
 
-    await new Promise((resolve, reject) => {
-      child.on('close', resolve);
-      child.on('error', reject);
-      child.on('exit', (exitCode) => {
-        if (exitCode !== 0) reject();
+  if (isTargetDirEmpty) {
+    console.log(`\n[kit-docs]: target directory is empty, creating new SvelteKit app.\n`);
+
+    try {
+      const child = spawn('npm', ['init', 'svelte@next', rootDirName ?? '.'], {
+        stdio: 'inherit',
       });
-    });
-  } catch (e) {
+
+      await new Promise((resolve, reject) => {
+        child.on('close', resolve);
+        child.on('error', reject);
+        child.on('exit', (exitCode) => {
+          if (exitCode !== 0) reject();
+        });
+      });
+    } catch (e) {
+      return;
+    }
+  }
+
+  if (!validateDirectory(targetDir)) {
     return;
   }
 
+  /** @type {RegExp[]} */
+  let overwrite = [/svelte\.config\.js/, /src\/routes\//];
+
+  if (!isTargetDirEmpty) {
+    /** @type {{ ok: boolean }} */
+    const { ok: svelteConfig } = await enquirer.prompt({
+      type: 'confirm',
+      name: 'ok',
+      message: `Overwrite \`svelte.config.js\`?`,
+      initial: false,
+    });
+
+    /** @type {{ ok: boolean }} */
+    const { ok: routes } = await enquirer.prompt({
+      type: 'confirm',
+      name: 'ok',
+      message: `Overwrite \`routes/\` directory?`,
+      initial: false,
+    });
+
+    overwrite = /** @type {RegExp[]} */ (
+      [svelteConfig && /svelte\.config\.js/, routes && /src\/routes\//].filter(Boolean)
+    );
+  }
+
   const pkgPath = path.resolve(targetDir, 'package.json');
-
-  while (!fs.existsSync(pkgPath));
-
   const pkg = JSON.parse(fs.readFileSync(pkgPath).toString());
 
-  pkg.devDependencies = {
+  const deps = {
     '@iconify-json/ri': '^1.1.1',
     '@sveltejs/adapter-auto': 'next',
     '@sveltejs/kit': 'next',
@@ -76,25 +95,44 @@ async function main() {
     svelte: '^3.44.0',
   };
 
+  if (!pkg.devDependencies) {
+    pkg.devDependencies = {};
+  }
+
+  // Add deps.
+  for (const dep of Object.keys(deps)) {
+    if (!pkg.dependencies?.[dep] && !pkg.devDependencies[dep]) {
+      pkg.devDependencies[dep] = deps[dep];
+    }
+  }
+
+  // Sort deps.
+  const sortedDeps = {};
+  for (const dep of Object.keys(pkg.devDependencies).sort()) {
+    sortedDeps[dep] = pkg.devDependencies[dep];
+  }
+  pkg.devDependencies = sortedDeps;
+
   fs.writeFileSync(pkgPath, JSON.stringify(pkg, null, 2) + '\n');
 
   const baseTemplateDir = path.resolve(__dirname, '../template-base');
 
-  copyDir(baseTemplateDir, targetDir);
+  copyDir(baseTemplateDir, targetDir, { overwrite });
 
   if (defaultTheme) {
     const themeTemplateDir = path.resolve(__dirname, '../template-theme');
-    copyDir(themeTemplateDir, targetDir);
+    copyDir(themeTemplateDir, targetDir, { overwrite });
   }
 
   const appDTSPath = path.resolve(targetDir, 'src/app.d.ts');
-  fs.writeFileSync(
-    appDTSPath,
-    fs
-      .readFileSync(appDTSPath)
-      .toString()
-      .replace(/\n/, `\n/// <reference types="@svelteness/kit-docs/globals" />\n`),
-  );
+  const appDTSContent = fs.readFileSync(appDTSPath).toString();
+
+  if (!/@svelteness\/kit-docs\/globals/.test(appDTSContent)) {
+    fs.writeFileSync(
+      appDTSPath,
+      appDTSContent.replace(/\n/, `\n/// <reference types="@svelteness/kit-docs/globals" />\n`),
+    );
+  }
 
   if (defaultTheme) {
     const appHTMLPath = path.resolve(targetDir, 'src/app.html');
@@ -112,46 +150,102 @@ async function main() {
     </script>
   `;
 
-    fs.writeFileSync(
-      appHTMLPath,
-      fs
-        .readFileSync(appHTMLPath)
-        .toString()
-        .replace(/%svelte.head%/, `${colorSchemeScript}\n    %svelte.head%`),
-    );
+    const fileContent = fs.readFileSync(appHTMLPath).toString();
+
+    if (!/svelteness::color-scheme/.test(fileContent)) {
+      fs.writeFileSync(
+        appHTMLPath,
+        fileContent.replace(/%svelte.head%/, `${colorSchemeScript}\n    %svelte.head%`),
+      );
+    }
   }
+
+  if (!isTargetDirEmpty && overwrite.length !== 2) {
+    console.log(
+      kleur.bold(
+        kleur.cyan(
+          `\n[kit-docs]: You will need to refer to the following manual install guides to finish setup:\n`,
+        ),
+      ),
+    );
+    console.log(
+      `\nhttps://kit-docs.svelteness.dev/docs/getting-started/quickstart#manual-installation`,
+    );
+    if (defaultTheme) {
+      console.log(
+        `\nhttps://kit-docs.svelteness.dev/docs/default-layout/installation#manual-installation`,
+      );
+    }
+  }
+
+  console.log();
 }
 
 /**
- * @returns {Promise<{ overwrite: boolean }>}
+ * @param {string} targetDir
+ * @returns {boolean}
  */
-function overwritePrompt() {
+function validateDirectory(targetDir) {
+  const requiredFiles = ['package.json', 'src/app.d.ts', 'src/app.html'];
+
+  for (const file of requiredFiles) {
+    const filePath = path.resolve(targetDir, 'package.json');
+    if (!fs.existsSync(filePath)) {
+      console.log(kleur.bold(kleur.red(`\n[kit-docs]: setup failed!`)));
+      console.log(`\nInvalid project directory, file is missing (${file}).\n`);
+      return false;
+    }
+  }
+
+  return true;
+}
+
+/**
+ * @returns {Promise<{ ok: boolean }>}
+ */
+function useCurrentDirPrompt() {
+  console.log(kleur.bold(kleur.cyan(`\n[kit-docs]: No directory was provided.\n`)));
+
   return enquirer.prompt({
     type: 'confirm',
-    name: 'overwrite',
-    message: `Directory not empty. Destroy?`,
+    name: 'ok',
+    message: `Use current directory (${process.cwd()})?`,
     initial: false,
   });
 }
 
 /**
- * @returns {Promise<{ defaultTheme: boolean }>}
+ * @returns {Promise<{ ok: boolean }>}
  */
 function defaultThemePrompt() {
   return enquirer.prompt({
     type: 'confirm',
-    name: 'defaultTheme',
+    name: 'ok',
     message: `Default KitDocs Theme?`,
     initial: true,
   });
 }
 
 /**
+ * @param {string} path
+ * @returns {boolean}
+ */
+function isDirEmpty(path) {
+  if (!fs.existsSync(path)) {
+    return true;
+  }
+
+  return fs.readdirSync(path).length === 0;
+}
+
+/**
  * @param {string} src
  * @param {string} dest
+ * @param {{ overwrite: RegExp[]}} options
  */
-function copyDir(src, dest) {
+function copyDir(src, dest, { overwrite }) {
   const filePaths = readDirDeepSync(src);
+
   for (const filePath of filePaths) {
     const destPath = path.resolve(dest, path.relative(src, filePath));
     const destDir = path.dirname(destPath);
@@ -160,7 +254,9 @@ function copyDir(src, dest) {
       fs.mkdirSync(destDir, { recursive: true });
     }
 
-    fs.writeFileSync(destPath, fs.readFileSync(filePath));
+    if (!fs.existsSync(filePath) || overwrite.some((o) => o.test(filePath))) {
+      fs.writeFileSync(destPath, fs.readFileSync(filePath));
+    }
   }
 }
